+++ packages/clerk-js/src/core/resources/AuthConfig.ts
@@ -5,6 +5,7 @@ import { BaseResource } from './internal';
 
 export class AuthConfig extends BaseResource implements AuthConfigResource {
   claimedAt: Date | null = null;
+  reverification: boolean = true;
   singleSessionMode: boolean = true;
 
   public constructor(data: AuthConfigJSON | null = null) {
@@ -19,17 +20,18 @@ export class AuthConfig extends BaseResource implements AuthConfigResource {
       return this;
     }
     this.claimedAt = data.claimed_at ? unixEpochToDate(data.claimed_at) : null;
+    this.reverification = data.reverification ?? true;
     this.singleSessionMode = data.single_session_mode ?? true;
-
     return this;
   }
 
   public __internal_toSnapshot(): AuthConfigJSONSnapshot {
     return {
       object: 'auth_config',
-      id: this.id ?? '',
-      single_session_mode: this.singleSessionMode,
       claimed_at: this.claimedAt ? this.claimedAt.getTime() : null,
+      id: this.id ?? '',
+      reverification: this.reverification,
+      single_session_mode: this.singleSessionMode,
     };
   }
 }