+++ packages/fastify/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.1.25
+
+### Patch Changes
+
+- Updated dependencies [[`67f1743aa1e0705d89ee6b532007f2686929240b`](https://github.com/clerk/javascript/commit/67f1743aa1e0705d89ee6b532007f2686929240b)]:
+  - @clerk/backend@1.25.1
+
 ## 2.1.24
 
 ### Patch Changes